requirements.txt
# Dependencias del proyecto ML-Ayram
# Instalar con: pip install -r requirements.txt
#
# NOTA DE COMPATIBILIDAD — Python 3.14.3:
# TensorFlow NO tiene wheel oficial para Python 3.14.
# Usamos PyTorch para la LSTM (misma funcionalidad, mejor soporte en versiones nuevas).
# Si alguna librería falla al instalar, prueba: pip install --pre NOMBRE_LIBRERÍA
#
# El servidor de producción (Linux) usa Python 3.11 donde todo instala sin problemas.

# ── Conexión a cTrader ──────────────────────────────────────────────────────
ctrader-open-api==0.9.2   # Última versión estable (la librería nunca llegó a 2.x)

# ── Datos y análisis ────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=2.0.0
ta>=0.11.0                # Reemplaza pandas-ta (sin dependencia de numba, compatible Python 3.14)
scipy>=1.11.0
# TA-Lib requiere instalación manual del .whl en Windows:
# Descarga de: https://github.com/cgohlke/talib-build/releases
# Instala con: pip install TA_Lib-0.x.x-cpXX-cpXX-win_amd64.whl

# ── Machine Learning ────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
torch>=2.2.0                 # PyTorch para LSTM (compatible Python 3.14)
# tensorflow NO tiene wheel para Python 3.14 — sustituido por PyTorch
optuna>=3.4.0
# mlflow>=2.9.0  # Solo en servidor (Python 3.11) — pyarrow no compila en Python 3.14 sin cmake+MSVC
joblib>=1.3.0

# ── Base de datos ───────────────────────────────────────────────────────────
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# ── Telegram Bot ────────────────────────────────────────────────────────────
python-telegram-bot>=21.0.0

# ── Scheduler y async ───────────────────────────────────────────────────────
APScheduler>=3.10.0

# ── Configuración y utilidades ──────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.0
rich>=13.0.0

# ── HTTP y APIs externas ─────────────────────────────────────────────────────
httpx>=0.26.0
requests>=2.31.0

# ── Desarrollo y análisis ────────────────────────────────────────────────────
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
